SAP: Update UI connection and notifications properly
This patch update connection state to upper layers only when SAP is
connected i.e. status indication is sent to SAP Client. Without this patch,
UI connected notification was triggered as soon as SAP rfcomm connection
was opened.
Also notification to disconnect SAP Server from UI is shown only when SAP
is successfully connected, and not when SapServer object is created.

Change-Id: Iaf40584220b51f6193bed9a54d159cffd32698e9