PBAP: Restart listeners if PBAP Service is not killed.
Usecase:
1) Perform BT off and on in quick succession on DUT.
2) Do SDP search from Remote device and check available services.
3) Repeat the above steps 10-15 times

Observation:
It is observed that PBAP Service cannot be found in SDP response of DUT

Root Cause:
During quick on-off scenarios, it is possible that On is queued after off,
so AdapterService is not killed by Frameworks, and this leads to PbapService
not getting killed. So when BT is turned on, listeners are not started, as
the code for starting listeners is present in OnCreate which is not called
because PBAP Service was never killed

Fix:
Add check in code to restart listeners on Adapter State change also, if they
have not been started already.

CRs-Fixed: 1020188
Change-Id: I0b13059991ff9d5526a181e55f6226e52dc72ef6